proj3/stackshack/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
pymysql==1.1.1       
cryptography==44.0.1 
werkzeug==3.0.6         
python-dotenv==1.0.0 
pytest
pytest-cov
pytest-html